ref #477: Move logic to a new listener (with lower priority)
