+++ src/CoreBundle/RequestType/AbstractRequestType.php
@@ -24,12 +24,4 @@ abstract class AbstractRequestType implements RequestTypeInterface
             header('Content-type: text/html; charset=UTF-8');
         }
     }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setAllowedDomains(array $domainNames): void
-    {
-
-    }
 }
