+++ src/CoreBundle/RequestType/FrontendRequestType.php
@@ -13,8 +13,6 @@ namespace ChameleonSystem\CoreBundle\RequestType;
 
 class FrontendRequestType extends AbstractRequestType
 {
-    private $allowedDomainNames = [];
-
     /**
      * @return int
      */
@@ -44,31 +42,12 @@ class FrontendRequestType extends AbstractRequestType
     protected function sendDefaultHeaders()
     {
         parent::sendDefaultHeaders();
-
-        $this->allowAdditionalDomains();
     }
-
-    private function allowAdditionalDomains(): void
-    {
-        foreach ($this->allowedDomainNames as $allowedDomain) {
-            // TODO some sort of "header encode"?
-            header("X-Frame-Options: ALLOW-FROM $allowedDomain");
-        }
-    }
-
     /**
      * {@inheritdoc}
      */
     public function getControllerName()
     {
         return 'chameleon_system_core.frontend_controller';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setAllowedDomains(array $domainNames): void
-    {
-        $this->allowedDomainNames = $domainNames;
     }
 }
